fix: normalize generic dispatch error type

+++ sim/coralnpu/hw_sim/gem5_bridge/gem5_generic_command_dispatch.cc
@@ -276,7 +276,8 @@ bool DispatchGem5GenericCommand(coral_operator_descriptor* descriptor,
   if (!ExecuteGem5FunctionalRequest(request, extmem, extmem_base,
                                     extmem_size)) {
     descriptor->error = request->error == CORAL_OPERATOR_ERROR_NONE
-                            ? CORAL_OPERATOR_ERROR_BAD_DESCRIPTOR
+                            ? static_cast<uint32_t>(
+                                  CORAL_OPERATOR_ERROR_BAD_DESCRIPTOR)
                             : request->error;
     return false;
   }
